+++ libpysal/io/fileio.py
