+++ libpysal/io/iohandlers/pyDbfIO.py
@@ -1,4 +1,4 @@
-from .. import Tables
+from .. import tables
 from ...common import MISSINGVALUE
 import datetime
 import struct
@@ -9,7 +9,7 @@ __author__ = "Charles R Schmidt <schmidtc@gmail.com>"
 __all__ = ['DBF']
 
 
-class DBF(Tables.DataTable):
+class DBF(tables.DataTable):
     """
     PySAL DBF Reader/Writer
 
@@ -51,7 +51,7 @@ class DBF(Tables.DataTable):
         dataPath -- str -- Path to file, including file.
         mode -- str -- 'r' or 'w'
         """
-        Tables.DataTable.__init__(self, *args, **kwargs)
+        tables.DataTable.__init__(self, *args, **kwargs)
         if self.mode == 'r':
             self.f = f = open(self.dataPath, 'rb')
             numrec, lenheader = struct.unpack('<xxxxLH22x', f.read(32)) #from dbf file standards
@@ -257,7 +257,7 @@ class DBF(Tables.DataTable):
             # End of file
             self.f.write('\x1A'.encode())
         self.f.close()
-        Tables.DataTable.close(self)
+        tables.DataTable.close(self)
 
     def _firstWrite(self, obj):
         if not self.header:
