+++ libpysal/io/iohandlers/template.py
