+++ libpysal/io/iohandlers/tests/test_arcgis_swm.py
@@ -1,15 +1,15 @@
 import unittest
-from ..wk1 import Wk1IO
+from ..arcgis_swm import ArcGISSwmIO
+from ...fileio import FileIO as psopen
 from .... import examples as pysal_examples
-from ...FileIO import FileIO as psopen
 import tempfile
 import os
 
 
-class test_Wk1IO(unittest.TestCase):
+class test_ArcGISSwmIO(unittest.TestCase):
     def setUp(self):
-        self.test_file = test_file = pysal_examples.get_path('spat-sym-us.wk1')
-        self.obj = Wk1IO(test_file, 'r')
+        self.test_file = test_file = pysal_examples.get_path('ohio.swm')
+        self.obj = ArcGISSwmIO(test_file, 'r')
 
     def test_close(self):
         f = self.obj
@@ -18,8 +18,8 @@ class test_Wk1IO(unittest.TestCase):
 
     def test_read(self):
         w = self.obj.read()
-        self.assertEqual(46, w.n)
-        self.assertEqual(4.0869565217391308, w.mean_neighbors)
+        self.assertEqual(88, w.n)
+        self.assertEqual(5.25, w.mean_neighbors)
         self.assertEqual([1.0, 1.0, 1.0, 1.0], list(w[1].values()))
 
     def test_seek(self):
@@ -31,7 +31,7 @@ class test_Wk1IO(unittest.TestCase):
     def test_write(self):
         w = self.obj.read()
         f = tempfile.NamedTemporaryFile(
-            suffix='.wk1', dir=pysal_examples.get_path(''))
+            suffix='.swm', dir=pysal_examples.get_path(''))
         fname = f.name
         f.close()
         o = psopen(fname, 'w')
