+++ libpysal/io/iohandlers/tests/test_arcgis_txt.py
@@ -1,16 +1,16 @@
 import unittest
-from ...FileIO import FileIO as psopen
+from ..arcgis_txt import ArcGISTextIO
+from ...fileio import FileIO as psopen
 from .... import examples as pysal_examples
-from ..arcgis_dbf import ArcGISDbfIO
 import tempfile
 import os
 import warnings
 
 
-class test_ArcGISDbfIO(unittest.TestCase):
+class test_ArcGISTextIO(unittest.TestCase):
     def setUp(self):
-        self.test_file = test_file = pysal_examples.get_path('arcgis_ohio.dbf')
-        self.obj = ArcGISDbfIO(test_file, 'r')
+        self.test_file = test_file = pysal_examples.get_path('arcgis_txt.txt')
+        self.obj = ArcGISTextIO(test_file, 'r')
 
     def test_close(self):
         f = self.obj
@@ -23,10 +23,10 @@ class test_ArcGISDbfIO(unittest.TestCase):
             w = self.obj.read()
             if len(warn) > 0:
                 assert issubclass(warn[0].category, RuntimeWarning)
-                assert "Missing Value Found, setting value to pysal.MISSINGVALUE" in str(warn[0].message)
-        self.assertEqual(88, w.n)
-        self.assertEqual(5.25, w.mean_neighbors)
-        self.assertEqual([1.0, 1.0, 1.0, 1.0], list(w[1].values()))
+                assert "DBF relating to ArcGIS TEXT was not found, proceeding with unordered string ids." in str(warn[0].message)
+        self.assertEqual(3, w.n)
+        self.assertEqual(2.0, w.mean_neighbors)
+        self.assertEqual([0.1, 0.05], list(w[2].values()))
 
     def test_seek(self):
         self.test_read()
@@ -40,17 +40,20 @@ class test_ArcGISDbfIO(unittest.TestCase):
             w = self.obj.read()
             if len(warn) > 0:
                 assert issubclass(warn[0].category, RuntimeWarning)
-                assert "Missing Value Found, setting value to pysal.MISSINGVALUE" in str(warn[0].message)
+                assert "DBF relating to ArcGIS TEXT was not found, proceeding with unordered string ids." in str(warn[0].message)
         f = tempfile.NamedTemporaryFile(
-            suffix='.dbf', dir=pysal_examples.get_path(''))
+            suffix='.txt', dir=pysal_examples.get_path(''))
         fname = f.name
         f.close()
-        o = psopen(fname, 'w', 'arcgis_dbf')
+        o = psopen(fname, 'w', 'arcgis_text')
         o.write(w)
         o.close()
-        f = psopen(fname, 'r', 'arcgis_dbf')
-        wnew = f.read()
-        f.close()
+        with warnings.catch_warnings(record=True) as warn:
+            warnings.simplefilter("always")
+            wnew = psopen(fname, 'r', 'arcgis_text').read()
+            if len(warn) > 0:
+                assert issubclass(warn[0].category, RuntimeWarning)
+                assert "DBF relating to ArcGIS TEXT was not found, proceeding with unordered string ids." in str(warn[0].message)
         self.assertEqual(wnew.pct_nonzero, w.pct_nonzero)
         os.remove(fname)
