+++ libpysal/io/iohandlers/tests/test_geobugs_txt.py
@@ -1,7 +1,7 @@
 import unittest
 from ..geobugs_txt import GeoBUGSTextIO
 from .... import examples as pysal_examples
-from ...FileIO import FileIO as psopen
+from ...fileio import FileIO as psopen
 import tempfile
 import os
