+++ libpysal/io/iohandlers/tests/test_stata_txt.py
@@ -1,7 +1,7 @@
 import unittest
 from .... import examples as pysal_examples
 from ..stata_txt import StataTextIO
-from ...FileIO import FileIO as psopen
+from ...fileio import FileIO as psopen
 import tempfile
 import os
