+++ libpysal/io/iohandlers/tests/test_wk1.py
@@ -1,16 +1,15 @@
 import unittest
-from ..mat import MatIO
+from ..wk1 import Wk1IO
 from .... import examples as pysal_examples
-from ...FileIO import FileIO as psopen
+from ...fileio import FileIO as psopen
 import tempfile
 import os
-import warnings
 
 
-class test_MatIO(unittest.TestCase):
+class test_Wk1IO(unittest.TestCase):
     def setUp(self):
-        self.test_file = test_file = pysal_examples.get_path('spat-sym-us.mat')
-        self.obj = MatIO(test_file, 'r')
+        self.test_file = test_file = pysal_examples.get_path('spat-sym-us.wk1')
+        self.obj = Wk1IO(test_file, 'r')
 
     def test_close(self):
         f = self.obj
@@ -32,15 +31,11 @@ class test_MatIO(unittest.TestCase):
     def test_write(self):
         w = self.obj.read()
         f = tempfile.NamedTemporaryFile(
-            suffix='.mat', dir=pysal_examples.get_path(''))
+            suffix='.wk1', dir=pysal_examples.get_path(''))
         fname = f.name
         f.close()
         o = psopen(fname, 'w')
-        with warnings.catch_warnings(record=True) as warn:
-            warnings.simplefilter("always")
-            o.write(w)
-            if len(warn) > 0:
-                assert issubclass(warn[0].category, FutureWarning)
+        o.write(w)
         o.close()
         wnew = psopen(fname, 'r').read()
         self.assertEqual(wnew.pct_nonzero, w.pct_nonzero)
