+++ libpysal/weights/util.py
@@ -1,4 +1,4 @@
-from ..io.FileIO import FileIO as psopen
+from ..io.fileio import FileIO as psopen
 from .weights import W, WSP
 from .Wsets import w_subset
 import numpy as np
